GPT-74 - Topology Layer - Specifically Fix WFS for ArcGis (ok on Geoserver)
* Fixed some failing tests I missed before inc. moving tests I wrote to an existing test file for same class
